+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add an User API endpoint to fisrt generate a JWT for Authentication purpose
   from Third Party Application
 
+### Changed
+
+- Generate JWT Token through Simple JWT's AccessToken class
+
 ## [0.2.1] - 2019-10-10
 
 ### Fixed